Derive S5 Modal Logic from HOOO EP
- Moved old "modal" module to "ava_modal"
- Added `hooo::tauto_to_para_not`
- Renamed "examples/modal" to "examples/ava_modal"
